requirements.txt
# Lodgeick Python Dependencies

# AI Integration
anthropic>=0.18.0

# Google Cloud APIs
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.115.0

# HTTP Requests
requests>=2.31.0

# Data Processing
python-dateutil>=2.8.2
